+++ CHANGES.rst
@@ -4,6 +4,10 @@ Changelog
 1.0dev (unreleased)
 -------------------
 
+- Re-Add cmf_edit method for ATEvent to ensure better backwards compatibility.
+  Move related cmf_edit tests from Products.CMFPlone to plone.app.event.
+  [thet]
+
 - Add Event profile definition for ATEvent completly in order to remove it from
   Products.CMFPlone. ATEvent is installed by ATContentTypes automatically as
   part of upcoming plone.app.contenttypes merge.